dÃebugg connexion mongodb pre-refacto

+++ src/main/java/com/hrizzon2/demotest/dao/FormationDao.java
@@ -8,5 +8,5 @@ import org.springframework.stereotype.Repository;
 public interface FormationDao extends JpaRepository<Formation, Integer> {
 
     // Exemple : compter stagiaires par formation
-    long countStagiairesById(Integer formationId);
+    int countStagiairesById(Integer formationId);
 }
